requirements.txt
appdirs==1.4.3
attrs==19.1.0
rply==0.7.7
